quickcheck tests for bigints
