+++ borsh/Cargo.toml
@@ -26,6 +26,9 @@ num-bigint = { version = "0.4.3", optional = true }
 borsh-derive = { path = "../borsh-derive" }
 hashbrown = "0.11"
 
+[dev-dependencies]
+quickcheck = "1.0"
+
 [features]
 default = ["std"]
 std = []
